textshift/backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.128.0"}
psycopg = {extras = ["binary"], version = "^3.3.2"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.22"
pydantic-settings = "^2.12.0"
torch = ">=2.0.0,<2.10.0"
transformers = "^5.0.0"
sentence-transformers = "^5.2.2"
redis = "^7.1.0"
sqlalchemy = "^2.0.46"
alembic = "^1.18.3"
boto3 = "^1.42.38"
httpx = "^0.28.1"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
